loja_web/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
jinja2
reportlab
python-barcode
Pillow
